+++ src/components/Header/Header.styles.ts
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	container: {
+		height: 48,
+		paddingHorizontal: 32,
+		flexDirection: 'row',
+		justifyContent: 'space-between',
+		alignItems: 'center'
+	},
+
+	title: {
+		flexDirection: 'row',
+		justifyContent: 'space-between',
+		alignItems: 'center',
+		marginBottom: 24,
+		paddingHorizontal: 24
+	}
+});
