+++ src/screens/DepositScreen/Deposit/Deposit.styles.ts
@@ -1,22 +1,6 @@
 import { StyleSheet } from 'react-native';
 
 export default StyleSheet.create({
-	header: {
-		height: 48,
-		paddingHorizontal: 32,
-		flexDirection: 'row',
-		justifyContent: 'space-between',
-		alignItems: 'center'
-	},
-
-	depositHeadline: {
-		flexDirection: 'row',
-		justifyContent: 'space-between',
-		alignItems: 'center',
-		marginBottom: 24,
-		paddingHorizontal: 24
-	},
-
 	depositButton: {
 		paddingHorizontal: 24,
 		marginTop: 'auto',
